functions/requirements.txt
fastapi
uvicorn
transformers
torch
python-multipart
pillow